pyproject.toml
[project]
name = "healthyhabitstrackerspa"
version = "0.1.0"
description = "SPA Web Application"
authors = [
    {name = "ahitahir@yandex.ru"}
]

license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.16.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "django (>=5.2.4,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "ipython (>=9.4.0,<10.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "eventlet (>=0.40.1,<0.41.0)",
    "redis (>=6.2.0,<7.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "coverage (>=7.9.2,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
